% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{add_association_rule}
\alias{add_association_rule}
\title{Add an association rule to the list of rules.}
\usage{
add_association_rule(
  rules,
  antecedent,
  consequence,
  support,
  confidence,
  fitness
)
}
\arguments{
\item{rules}{The current list of association rules.}

\item{antecedent}{The antecedent part of the association rule.}

\item{consequence}{The consequent part of the association rule.}

\item{support}{The support of the association rule.}

\item{confidence}{The confidence of the association rule.}

\item{fitness}{The fitness of the association rule.}
}
\value{
The updated list of association rules.
}
\description{
This function adds a new association rule to the existing list of rules.
}
