% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{add_attribute}
\alias{add_attribute}
\title{Add an attribute to the "rule" list.}
\usage{
add_attribute(rules, name, type, border1, border2, value)
}
\arguments{
\item{rules}{The current rules list.}

\item{name}{The name of the feature in the rule.}

\item{type}{The type of the feature in the rule.}

\item{border1}{The first border value in the rule.}

\item{border2}{The second border value in the rule.}

\item{value}{The value associated with the rule.}
}
\value{
The updated rules list.
}
\description{
This function adds an attribute to the existing list.
}
\examples{
rules <- list()
new_rules <- add_attribute(rules, "feature1", "numerical", 0.2, 0.8, "EMPTY")

}
