% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_layout_cpp}
\alias{build_layout_cpp}
\title{Entry point for R to generate coral plot data from a set of association rules.}
\usage{
build_layout_cpp(parsed, grid_size, lhs_sort = "confidence")
}
\arguments{
\item{parsed}{TODO}

\item{grid_size}{Number of grid cells per layout dimension to use during node positioning.}

\item{lhs_sort}{The sorting metric for tzhe antecedent}
}
\value{
A List containing two DataFrames:
\itemize{
   \item \code{edges}: DataFrame with start and end coordinates, line width, and color for each edge.
   \item \code{nodes}: DataFrame with position, radius, ID, and item label for each node.
}
}
\description{
This function takes a list of parsed association rules and produces two data frames:
one for the nodes and one for the edges of a coral plot. It acts as a wrapper that:
\itemize{
   \item Constructs the coral layout.
   \item Converts the resulting nodes and edges into R-compatible data frames.
}
}
\keyword{internal}
