% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_radial_plots.R
\name{build_radial_plots}
\alias{build_radial_plots}
\title{Build radial‐plot data directly from niarules output}
\usage{
build_radial_plots(arules)
}
\arguments{
\item{arules}{TODO}
}
\value{
TODO
}
\description{
TODO: full description here
}
