% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{build_rule}
\alias{build_rule}
\title{Build rules based on a candidate solution.}
\usage{
build_rule(solution, features)
}
\arguments{
\item{solution}{The solution vector.}

\item{features}{The features list.}
}
\value{
A rule.
}
\description{
This function takes a candidate solution vector and a features list and builds rule.
}
