% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{calculate_fitness}
\alias{calculate_fitness}
\title{Calculate the fitness of an association rule.}
\usage{
calculate_fitness(supp, conf)
}
\arguments{
\item{supp}{The support of the association rule.}

\item{conf}{The confidence of the association rule.}
}
\value{
The fitness of the association rule.
}
\description{
This function calculates the fitness of an association rule using support and confidence.
}
