% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{calculate_selected_category}
\alias{calculate_selected_category}
\title{Calculate the selected category based on a value and the number of categories.}
\usage{
calculate_selected_category(value, num_categories)
}
\arguments{
\item{value}{The value to calculate the category for.}

\item{num_categories}{The total number of categories.}
}
\value{
The calculated selected category.
}
\description{
This function calculates the selected category based on a given value and
the total number of categories.
}
\examples{
selected_category <- calculate_selected_category(0.3, 5)

}
