% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{check_attribute}
\alias{check_attribute}
\title{Check if the attribute conditions are satisfied for an instance.}
\usage{
check_attribute(attribute, instance_row)
}
\arguments{
\item{attribute}{An attribute with type and name information.}

\item{instance_row}{A row representing an instance in the dataset.}
}
\value{
TRUE if conditions are satisfied, FALSE otherwise.
}
\description{
This function checks if the attribute conditions specified in the association rule
are satisfied for a given instance row.
}
