% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_theme.R
\name{coral_get_theme}
\alias{coral_get_theme}
\title{Get a theme object (experimental)}
\usage{
coral_get_theme(
  name = c("default", "studio", "flat", "dark", "none"),
  overrides = NULL
)
}
\arguments{
\item{name}{one of coral_list_themes()}

\item{overrides}{named list to tweak values}
}
\value{
an object of class 'coral_theme'
}
\description{
`r lifecycle::badge('experimental')`
}
