% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_theme.R
\name{coral_list_themes}
\alias{coral_list_themes}
\title{Available themes (experimental)}
\usage{
coral_list_themes()
}
\value{
character vector of theme names
}
\description{
`r lifecycle::badge('experimental')`
}
