% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{cut_point}
\alias{cut_point}
\title{Calculate the cut point for an association rule.}
\usage{
cut_point(sol, num_attr)
}
\arguments{
\item{sol}{The cut value from the solution vector.}

\item{num_attr}{The number of attributes in the association rule.}
}
\value{
The cut point value.
}
\description{
This function calculates the cut point, denoting which part of the vector belongs
to the antecedent and which to the consequent of the mined association rule.
}
