% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a candidate solution.}
\usage{
evaluate(solution, features, instances)
}
\arguments{
\item{solution}{A vector representing a candidate solution for the association rule.}

\item{features}{A list containing information about features, including type and bounds.}

\item{instances}{A data frame representing instances in the dataset.}
}
\value{
The fitness of the association rule and identified rule.
}
\description{
This function takes a canditate solution (vector), list of features, and instances, and evaluates the fitness
of an association rule by calculating support and confidence.
}
