% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{feature_borders}
\alias{feature_borders}
\title{Get the lower and upper bounds of a feature.}
\usage{
feature_borders(features, name)
}
\arguments{
\item{features}{A list containing information about features, including type and bounds.}

\item{name}{The name of the feature.}
}
\value{
A list containing the lower and upper bounds of the feature.
}
\description{
This function retrieves the lower and upper bounds of a feature from the features list.
}
