% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{feature_position}
\alias{feature_position}
\title{Get the position of a feature.}
\usage{
feature_position(features, feature)
}
\arguments{
\item{features}{The features list.}

\item{feature}{The name of the feature to find.}
}
\value{
The position of the feature.
}
\description{
This function returns the position of a feature in the vector,
considering the type of the feature.
}
\examples{
features <- list(
  feature1 = list(type = "numerical"),
  feature2 = list(type = "categorical"),
  feature3 = list(type = "numerical")
)
position <- feature_position(features, "feature2")

}
