% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_helpers.R
\name{filter_parsed_by_rhs}
\alias{filter_parsed_by_rhs}
\title{Filter parsed object to a single RHS and remap item ids to 0..K-1}
\usage{
filter_parsed_by_rhs(parsed, rhs_label, ignore_order = TRUE)
}
\description{
Filter parsed object to a single RHS and remap item ids to 0..K-1
}
