% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format_rule_parts}
\alias{format_rule_parts}
\title{Format Rule Parts}
\usage{
format_rule_parts(parts)
}
\arguments{
\item{parts}{A list containing parts of an association rule.}
}
\value{
A formatted string representing the rule parts.
}
\description{
This function formats the parts of an association rule into a string.
}
