% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_helpers.R
\name{make_node_color_map}
\alias{make_node_color_map}
\title{Stable node color map (HCL)}
\usage{
make_node_color_map(parsed, by = c("type", "item"), hcl_c = 65, hcl_l = 70)
}
\description{
Stable node color map (HCL)
}
