% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print_association_rules}
\alias{print_association_rules}
\title{Print Numerical Association Rules}
\usage{
print_association_rules(rules)
}
\arguments{
\item{rules}{A list containing association rules.}
}
\value{
Prints the association rules.
}
\description{
This function prints association rules including antecedent, consequence,
support, confidence, and fitness.
}
