% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print_rule_parts}
\alias{print_rule_parts}
\title{Print Rule Parts}
\usage{
print_rule_parts(parts)
}
\arguments{
\item{parts}{A list containing parts of an association rule.}
}
\value{
Prints the rule parts.
}
\description{
This function prints the parts of an association rule, including name, type,
border1, border2, and value.
}
