% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{problem_dimension}
\alias{problem_dimension}
\title{Calculate the dimension of the problem, excluding timestamps.}
\usage{
problem_dimension(feature_info, is_time_series = FALSE)
}
\arguments{
\item{feature_info}{A list containing information about each feature.}

\item{is_time_series}{Boolean indicating if time series data is present.}
}
\value{
The calculated dimension based on the feature types.
}
\description{
Calculate the dimension of the problem, excluding timestamps.
}
