% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Read a dataset from a CSV file.}
\usage{
read_dataset(dataset_path)
}
\arguments{
\item{dataset_path}{The path to the CSV file.}
}
\value{
A data frame representing the dataset.
}
\description{
This function reads a CSV file and returns the dataset as a data frame.
}
\examples{
read_dataset("Abalone.csv")

}
