% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_plot_visualization.R
\name{render_radial_rgl}
\alias{render_radial_rgl}
\title{Render radial plots in 3D}
\usage{
render_radial_rgl(
  nodes,
  edges,
  grid_size,
  feature_cols = NULL,
  grid_color = "grey80",
  legend = FALSE
)
}
\arguments{
\item{nodes}{TODO}

\item{edges}{TODO}

\item{grid_size}{TODO}

\item{feature_cols}{TODO}

\item{grid_color}{TODO}

\item{legend}{TODO}
}
\description{
TODO: full description here
}
