% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{rs}
\alias{rs}
\title{Simple Random Search}
\usage{
rs(candidate_len)
}
\arguments{
\item{candidate_len}{The length of the vector of random solutions.}
}
\value{
A vector of random solutions between 0 and 1.
}
\description{
This function generates a vector of random solutions for a specified length.
}
\examples{
\dontrun{
candidate_len <- 10
random_solutions <- rs(candidate_len)
print(random_solutions)
}

}
