% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{supp_conf}
\alias{supp_conf}
\title{Calculate support and confidence for an association rule.}
\usage{
supp_conf(antecedent, consequent, instances, features)
}
\arguments{
\item{antecedent}{The antecedent part of the association rule.}

\item{consequent}{The consequent part of the association rule.}

\item{instances}{A data frame representing instances in the dataset.}

\item{features}{A list containing information about features, including type and bounds.}
}
\value{
A list containing support and confidence values.
}
\description{
This function calculates the support and confidence for the given antecedent
and consequent in the dataset instances.
}
