% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_association_rules_to_csv}
\alias{write_association_rules_to_csv}
\title{Write Association Rules to CSV file}
\usage{
write_association_rules_to_csv(rules, file_path)
}
\arguments{
\item{rules}{A list of association rules.}

\item{file_path}{The file path for the CSV output.}
}
\description{
This function writes association rules to a CSV file.
}
