% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{extract_feature_info}
\alias{extract_feature_info}
\title{Extract feature information from a dataset.}
\usage{
extract_feature_info(data)
}
\arguments{
\item{data}{The dataset to analyze.}
}
\value{
A list containing information about each feature.
}
\description{
This function analyzes the given dataset and extracts information about each feature.
}
