% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{map_to_ts}
\alias{map_to_ts}
\title{Map solution boundaries to time series indices.}
\usage{
map_to_ts(lower, upper, total_transactions)
}
\arguments{
\item{lower}{The lower bound of the time range in [0, 1].}

\item{upper}{The upper bound of the time range in [0, 1].}

\item{total_transactions}{The total number of transactions in the time series.}
}
\value{
A list with `low` and `up` indices.
}
\description{
This function maps the lower and upper bounds of the solution vector to
transaction indices based on the length of the time series.
}
