% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{print_feature_info}
\alias{print_feature_info}
\title{Print feature information extracted from a dataset.}
\usage{
print_feature_info(feature_info)
}
\arguments{
\item{feature_info}{The list containing information about each feature.}
}
\description{
This function prints the information extracted about each feature.
}
