% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Read a CSV Dataset}
\usage{
read_dataset(
  dataset_path,
  timestamp_col = "timestamp",
  timestamp_formats = c("\%d/\%m/\%Y \%H:\%M:\%S", "\%H:\%M:\%S \%d/\%m/\%Y")
)
}
\arguments{
\item{dataset_path}{A string specifying the path to the CSV file.}

\item{timestamp_col}{A string specifying the timestamp column name (default: `"timestamp"`).}

\item{timestamp_formats}{A vector of date-time formats to try for parsing timestamps.}
}
\value{
A data frame containing the dataset.
}
\description{
Reads a dataset from a CSV file and optionally parses a timestamp column.
}
