% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_plot_visualization.R
\name{render_coral_rgl}
\alias{render_coral_rgl}
\title{applies styling to coral plots and renders them using rgl}
\usage{
render_coral_rgl(
  nodes,
  edges,
  grid_size,
  grid_color = "grey80",
  legend = FALSE,
  label_mode = c("none", "interval", "item", "interval_short"),
  label_cex = 0.7,
  label_offset = 1.5,
  max_labels = 100,
  edge_width_metric = c("confidence", "lift", "support"),
  edge_color_metric = c("confidence", "lift", "support"),
  edge_alpha_metric = NULL,
  edge_width_range = c(1, 5),
  edge_width_transform = c("linear", "sqrt", "log"),
  edge_gradient = c("#2166AC", "#67A9CF", "#D1E5F0", "#FDDBC7", "#EF8A62", "#B2182B"),
  edge_color_transform = c("linear", "sqrt", "log"),
  edge_alpha = 0.6,
  edge_alpha_range = c(0.25, 0.5),
  edge_alpha_transform = c("linear", "sqrt", "log"),
  node_color_by = c("type", "item", "none", "edge_incoming", "edge_outgoing_mean"),
  node_gradient = "match",
  node_gradient_map = c("even", "hash", "frequency"),
  y_scale = 0,
  jitter_sd = 0,
  jitter_mode = c("deterministic", "random"),
  jitter_seed = NULL,
  return_data = FALSE
)
}
\arguments{
\item{nodes}{data.frame from build_coral_layout()$nodes}

\item{edges}{data.frame from build_coral_layout()$edges}

\item{grid_size}{integer from build_coral_layout()$grid_size}

\item{grid_color}{background grid color}

\item{legend}{show a node legend (by base feature)}

\item{label_mode}{"none"|"interval"|"item"|"interval_short"}

\item{label_cex}{label size}

\item{label_offset}{vertical offset in radii}

\item{max_labels}{max non-root labels to be shown}

\item{edge_width_metric}{which metric to map to edge width}

\item{edge_color_metric}{which metric to map to edge width}

\item{edge_alpha_metric}{which metric to map to edge width}

\item{edge_width_range}{numeric length-2, min/max lwd}

\item{edge_width_transform}{"linear"|"sqrt"|"log"}

\item{edge_gradient}{character vector (>=2) for edge colors}

\item{edge_color_transform}{"linear"|"sqrt"|"log"}

\item{edge_alpha}{edge transparency (0..1)}

\item{edge_alpha_range}{TODO}

\item{edge_alpha_transform}{"linear"|"sqrt"|"log"}

\item{node_color_by}{TODO}

\item{node_gradient}{TODO}

\item{node_gradient_map}{TODO}

\item{y_scale}{numeric scalar; vertical scale (0 = flat, default).}

\item{jitter_sd}{TODO}

\item{jitter_mode}{TODO}

\item{jitter_seed}{TODO,}

\item{return_data}{used for testing only}
}
\description{
Expects nodes/edges from build_coral_layout(). Edges must contain
columns: x,y,z,x_end,y_end,z_end and metrics: support, confidence, lift.
}
