% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_helpers.R
\name{rhs_label_per_rule}
\alias{rhs_label_per_rule}
\title{RHS label per rule (aligned vector)}
\usage{
rhs_label_per_rule(parsed, ignore_order = TRUE)
}
\description{
RHS label per rule (aligned vector)
}
