% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{write_association_rules_to_csv}
\alias{write_association_rules_to_csv}
\title{Write Association Rules to CSV file}
\usage{
write_association_rules_to_csv(rules, file_path)
}
\arguments{
\item{rules}{A list of association rules.}

\item{file_path}{The file path for the CSV output.}
}
\description{
This function writes association rules to a CSV file.
}
\examples{
\donttest{arules <- list(list(
  list(
    antecedent = list(
      list(name = "Diameter", type = "numerical", border1 = 0, border2 = 1, value = "EMPTY")
    ),
    consequence = list(
      list(name = "Height", type = "numerical", border1 = 0, border2 = 1, value = "EMPTY")
    ),
    support = 0.9997606,
    confidence = 0.9997606,
    fitness = 1.999521
  )
))
write_association_rules_to_csv(arules, "association_rules_output.csv")}

}
